camera-frame-capture/requirements.txt
loguru==0.7.2
numpy==1.26.1
opencv-python==4.8.1.78
python-dotenv==1.0.0
redis==5.0.1
